Refactors Crypto key service into separate key provider and parameter provider
