+++ Gerulus.Core/Crypto/ICryptoKeyProvider.cs
@@ -1,0 +1,7 @@
+namespace Gerulus.Core.Crypto;
+
+public interface ICryptoKeyProvider
+{
+    Task<CryptographicKeyPair> GenerateKeyPairAsync();
+    Task<byte[]> ComputeSharedKeyAsync(CryptographicKeyPair initiator, CryptographicKeyPair otherKey);
+}
