+++ Gerulus.Core/Crypto/ICryptoParameterProvider.cs
@@ -1,0 +1,10 @@
+namespace Gerulus.Core.Crypto;
+
+public interface ICryptoParameterProvider<TParameters>
+{
+    bool IsInitialized { get; }
+    TParameters? Parameters { get; }
+
+    Task<TParameters> GenerateParametersAsync();
+    Task InitializeParametersAsync(TParameters parameters);
+}
